Add an option to disable robot detection
In some cases, like Turtlebots with an RPLIDAR on top of them, robots
can't detect each other using their LIDAR. Since MRGraphSLAM only
considers that data match if other robots are seen, this is an issue.
This commit adds the -detectRobots option to disable this checking (it
remains on by default).